fix: Change read_count type to uint64 for syscall consistency
- Update read_count type from int to uint64 in defs.h and main.c
- Match syscall return type in sys_getreadcount()
- Remove redundant extern declarations in sysfile.c

+++ kernel/defs.h
@@ -146,7 +146,7 @@ int             fetchaddr(uint64, uint64*);
 void            syscall();
 
 // Global counter for tracking read() system call invocations
-extern int read_count;  // Counter for read() syscall tracking since boot
+extern uint64 read_count;  // Counter for read() syscall tracking since boot
 
 // trap.c
 extern uint     ticks;
